+++ src/context.rs
@@ -89,10 +89,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     action::fns::ActionFns,
-    binding::{
-        FirstActivation,
-        source::{BindingSourceFns, BindingSourceRegistry},
-    },
+    binding::FirstActivation,
     condition::fns::{ConditionFns, ConditionRegistry},
     context::{input_reader::PendingBindings, trigger_tracker::TriggerTracker},
     modifier::fns::{ModifierFns, ModifierRegistry},
@@ -209,7 +206,7 @@ pub(crate) struct ScheduleContexts {
     activity_ids: Vec<ComponentId>,
 
     /// Configures the app for this schedule.
-    setup: fn(&Self, &mut App, &ConditionRegistry, &ModifierRegistry, &BindingSourceRegistry),
+    setup: fn(&Self, &mut App, &ConditionRegistry, &ModifierRegistry),
 }
 
 impl ScheduleContexts {
@@ -234,9 +231,8 @@ impl ScheduleContexts {
         app: &mut App,
         conditions: &ConditionRegistry,
         modifiers: &ModifierRegistry,
-        sources: &BindingSourceRegistry,
     ) {
-        (self.setup)(self, app, conditions, modifiers, sources);
+        (self.setup)(self, app, conditions, modifiers);
     }
 
     /// Configures the app for all contexts registered for schedule `C`.
@@ -245,7 +241,6 @@ impl ScheduleContexts {
         app: &mut App,
         conditions: &ConditionRegistry,
         modifiers: &ModifierRegistry,
-        sources: &BindingSourceRegistry,
     ) {
         debug!("setting up systems for `{}`", ShortName::of::<S>());
 
@@ -275,9 +270,6 @@ impl ScheduleContexts {
                         builder.mut_id(id);
                     }
                     for &id in &**modifiers {
-                        builder.mut_id(id);
-                    }
-                    for &id in &**sources {
                         builder.mut_id(id);
                     }
                 });
@@ -458,7 +450,6 @@ fn update<S: ScheduleLabel>(
             &mut FirstActivation,
             Option<&ModifierFns>,
             Option<&ConditionFns>,
-            Option<&BindingSourceFns>,
         ),
         Without<ActionSettings>,
     >,
@@ -555,23 +546,11 @@ fn update<S: ScheduleLabel>(
                     mut first_activation,
                     modifiers,
                     conditions,
-                    source_fns,
                 )) = bindings_iter.fetch_next()
                 {
                     let mut binding_entity = conds_and_mods.get_mut(binding_entity).unwrap();
 
-                    let new_value = if let Binding::Custom = binding {
-                        let mut value = ActionValue::zero(dim);
-                        if let Some(source_fns) = source_fns {
-                            for get_source in &**source_fns {
-                                let source = get_source(&mut binding_entity);
-                                value = source.read(&actions_data, &time);
-                            }
-                        }
-                        value
-                    } else {
-                        reader.value(binding)
-                    };
+                    let new_value = reader.value(binding);
 
                     if action_settings.require_reset && **first_activation {
                         // Ignore until we read zero for this mapping.